build: add provenance and CI based builds
We added a quick workflow to build the package and also publish it to npm.

Publishing can then also utilize the provenance feature, to verify that the package was actually built in GitHub Actions on that exact commit. This will increase supply-chain security and trust.

We will now need to supply a npm API token via GitHub secrets, called `NPM_TOKEN`.
The workflow now triggers on release published.
So you can just create a tag and release on GitHub to trigger the action.

See https://github.blog/security/supply-chain-security/introducing-npm-package-provenance/